+++ ennoia/__init__.py
@@ -33,4 +33,4 @@ __all__ = [
     "describe",
 ]
 
-__version__ = "0.4.1"
+__version__ = "0.4.2"